requirements.txt
blabel==0.1.6
py-cord==2.5.0
python-dotenv==1.0.1
pillow==10.4.0
PyYAML==6.0.1
pyusb==1.2.1
pymupdf==1.25.4
requests==2.32.3
aiosqlite==0.20.0
aiohttp==3.11.11
aiofiles==24.1.0
memory-tempfile==2.2.3
git+https://github.com/marcopolo42/brother_ql

